requirements.txt
pytest~=8.4.1
requests~=2.32.4
pytest-html
allure-pytest
allure-python-commons~=2.15.0